// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
      books: {
        Row: {
          autori: string | null
          created_at: string
          data_fine: string | null
          data_inizio: string | null
          anno: number | null
          categoria: string | null
          editore: string | null
          id: number
          is_prestato: boolean
          isbn: string | null
          posizione: string | null
          prestato_a: string | null
          prestato_telefono: string | null
          titolo: string
          updated_at: string
        }
        Insert: {
          autori?: string | null
          created_at?: string
          data_fine?: string | null
          data_inizio?: string | null
          anno?: number | null
          categoria?: string | null
          editore?: string | null
          id?: never
          is_prestato?: boolean
          isbn?: string | null
          posizione?: string | null
          prestato_a?: string | null
          prestato_telefono?: string | null
          titolo: string
          updated_at?: string
        }
        Update: {
          autori?: string | null
          created_at?: string
          data_fine?: string | null
          data_inizio?: string | null
          anno?: number | null
          categoria?: string | null
          editore?: string | null
          id?: number
          is_prestato?: boolean
          isbn?: string | null
          posizione?: string | null
          prestato_a?: string | null
          prestato_telefono?: string | null
          titolo?: string
          updated_at?: string
        }
        Relationships: []
      }
      loan_history: {
        Row: {
          book_id: number
          created_at: string
          data_fine: string | null
          data_inizio: string
          id: number
          prestato_a: string
          prestato_telefono: string | null
        }
        Insert: {
          book_id: number
          created_at?: string
          data_fine?: string | null
          data_inizio: string
          id?: string
          prestato_a: string
          prestato_telefono?: string | null
        }
        Update: {
          book_id?: number
          created_at?: string
          data_fine?: string | null
          data_inizio?: string
          id?: number
          prestato_a?: string
          prestato_telefono?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "loan_history_book_id_fkey"
            columns: ["book_id"]
            isOneToOne: false
            referencedRelation: "books"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never